Merged in remove-mysql-sqlite-support (pull request #1)
Remove MySQL and SQLite support

This is a drastic measure, but 1) I don't use these DBs currently, 2) I
don't think anyone else is relying on this fork, 3) an upstream change
in one of them is breaking goose on OSX for me.

The more permanent fix is to vendor and pin all dependencies, but this
is a quick fix to get my other work unblocked.

Approved-by: jeffwatts <jrwatts@gmail.com>